Terrod spirte & scaling
Scaling Main, Level Selection and their canvases to 1920 x 1080
Had to delete some brackets and place them again becuase they were under the background. Sorting the layers wasn't working, but in the build and run works. WE GOOD
Added the cool down and match to their proper holder. There's no script for it.